add qc check for fmriprep output

+++ run_fmriprep_kids_alicja.sh
@@ -15,12 +15,13 @@ subj=${1##*-}
 work=$(realpath ~/fmriprep)
 
 mkdir -p $work/$subj
+mkdir -p ~/freesurfer-subjects-dir #if not exists will give error
 
 nprocs=6
 mem=10000 #mb
 
 export ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS=$nprocs
-fmriprep $root_dir \
+echo fmriprep $root_dir \
          $output_dir \
          participant \
          --fs-license-file ~/freesurfer.txt \
@@ -35,4 +36,19 @@ fmriprep $root_dir \
          --output-spaces MNIPediatricAsym:cohort-2:res-2 \
          -v
 
-rm -rf $work/$subj
+# ----- QC CHECK -----
+
+logfile="$output_dir/fmriprep.log"
+
+if find "$output_dir/sub-$subj" -type f -name "*desc-preproc_bold.nii.gz" -print -quit | grep -q .; then
+    if ! find "$root_dir/sub-$subj" -type d -name fmap | grep -q .; then
+        echo "$(date) sub-$subj : BOLD present but fmap missing" >> "$logfile"
+    else
+        echo "$(date) sub-$subj : Preproc finished" >> "$logfile"
+        rm -rf $work/$subj
+    fi
+
+else
+    echo "$(date) sub-$subj : Preproc failed (no desc-preproc_bold found)" >> "$logfile"
+fi
+
